Ensuring that errors raised while increment processed or notifying progress are silenced
When processing we could have a situation where an error is raised after incrementing the progress. This could happen when notifying for example. In this case when retrying we could end up having the job processed count incremented N number of times as there were retries. This change prevents that from happening by silencing exceptions raised on these methods.

This means that we may lose a progress increment or notification. The notification is optional, and it is okay to lose some of them. If we lose a progress increment than the final processed count would not be accurate. Since when mongoid increments the processed it also does it in memory when finishing it will also update the processed if necessary.